Disentangle data model tests by moving fixtures into their own file
The data model tests all had linker dependencies on each other, which meant
that running any of Test{Read,Write,Commands} would run the tests from all of
them, because the global fixture function from that test was required.

This change moves all those fixture functions into DataModelFixtures.cpp.
Also disambiguate the enum types and control variable names involved.
