+++ src/controller/tests/data_model/BUILD.gn
@@ -22,11 +22,19 @@ import("${chip_root}/src/platform/device.gni")
 chip_test_suite("data_model") {
   output_name = "libDataModelTests"
 
+  sources = [
+    "DataModelFixtures.cpp",
+    "DataModelFixtures.h",
+  ]
+
+  test_sources = []
   if (chip_device_platform != "mbed" && chip_device_platform != "efr32" &&
       chip_device_platform != "esp32" && chip_device_platform != "fake") {
-    test_sources = [ "TestCommands.cpp" ]
-    test_sources += [ "TestWrite.cpp" ]
-    test_sources += [ "TestRead.cpp" ]
+    test_sources += [
+      "TestCommands.cpp",
+      "TestRead.cpp",
+      "TestWrite.cpp",
+    ]
   }
 
   public_deps = [
