+++ src/controller/tests/data_model/TestCommands.cpp
@@ -24,10 +24,12 @@
 
 #include <gtest/gtest.h>
 
-#include "app/data-model/NullObject.h"
+#include "DataModelFixtures.h"
+
 #include <app-common/zap-generated/cluster-objects.h>
 #include <app/AppConfig.h>
 #include <app/InteractionModelEngine.h>
+#include <app/data-model/NullObject.h>
 #include <app/tests/AppTestContext.h>
 #include <controller/InvokeInteraction.h>
 #include <lib/core/CHIPCore.h>
@@ -44,141 +46,8 @@ using TestContext = chip::Test::AppContext;
 using namespace chip;
 using namespace chip::app;
 using namespace chip::app::Clusters;
+using namespace chip::app::DataModelTests;
 using namespace chip::Protocols;
-
-namespace {
-chip::ClusterStatus kTestSuccessClusterStatus = 1;
-chip::ClusterStatus kTestFailureClusterStatus = 2;
-
-constexpr EndpointId kTestEndpointId = 1;
-
-enum ResponseDirective
-{
-    kSendDataResponse,
-    kSendSuccessStatusCode,
-    kSendMultipleSuccessStatusCodes,
-    kSendError,
-    kSendMultipleErrors,
-    kSendSuccessStatusCodeWithClusterStatus,
-    kSendErrorWithClusterStatus,
-    kAsync,
-};
-
-ResponseDirective responseDirective;
-CommandHandler::Handle asyncHandle;
-
-} // namespace
-
-namespace chip {
-namespace app {
-
-void DispatchSingleClusterCommand(const ConcreteCommandPath & aCommandPath, chip::TLV::TLVReader & aReader,
-                                  CommandHandler * apCommandObj)
-{
-    ChipLogDetail(Controller, "Received Cluster Command: Endpoint=%x Cluster=" ChipLogFormatMEI " Command=" ChipLogFormatMEI,
-                  aCommandPath.mEndpointId, ChipLogValueMEI(aCommandPath.mClusterId), ChipLogValueMEI(aCommandPath.mCommandId));
-
-    if (aCommandPath.mClusterId == Clusters::UnitTesting::Id &&
-        aCommandPath.mCommandId == Clusters::UnitTesting::Commands::TestSimpleArgumentRequest::Type::GetCommandId())
-    {
-        Clusters::UnitTesting::Commands::TestSimpleArgumentRequest::DecodableType dataRequest;
-
-        if (DataModel::Decode(aReader, dataRequest) != CHIP_NO_ERROR)
-        {
-            apCommandObj->AddStatus(aCommandPath, Protocols::InteractionModel::Status::Failure, "Unable to decode the request");
-            return;
-        }
-
-        if (responseDirective == kSendDataResponse)
-        {
-            Clusters::UnitTesting::Commands::TestStructArrayArgumentResponse::Type dataResponse;
-            Clusters::UnitTesting::Structs::NestedStructList::Type nestedStructList[4];
-
-            uint8_t i = 0;
-            for (auto & item : nestedStructList)
-            {
-                item.a   = i;
-                item.b   = false;
-                item.c.a = i;
-                item.c.b = true;
-                i++;
-            }
-
-            dataResponse.arg1 = nestedStructList;
-            dataResponse.arg6 = true;
-
-            apCommandObj->AddResponse(aCommandPath, dataResponse);
-        }
-        else if (responseDirective == kSendSuccessStatusCode)
-        {
-            apCommandObj->AddStatus(aCommandPath, Protocols::InteractionModel::Status::Success);
-        }
-        else if (responseDirective == kSendMultipleSuccessStatusCodes)
-        {
-            apCommandObj->AddStatus(aCommandPath, Protocols::InteractionModel::Status::Success,
-                                    "No error but testing status success case");
-
-            // TODO: Right now all but the first AddStatus call fail, so this
-            // test is not really testing what it should.
-            for (size_t i = 0; i < 3; ++i)
-            {
-                (void) apCommandObj->FallibleAddStatus(aCommandPath, Protocols::InteractionModel::Status::Success,
-                                                       "No error but testing status success case");
-            }
-            // And one failure on the end.
-            (void) apCommandObj->FallibleAddStatus(aCommandPath, Protocols::InteractionModel::Status::Failure);
-        }
-        else if (responseDirective == kSendError)
-        {
-            apCommandObj->AddStatus(aCommandPath, Protocols::InteractionModel::Status::Failure);
-        }
-        else if (responseDirective == kSendMultipleErrors)
-        {
-            apCommandObj->AddStatus(aCommandPath, Protocols::InteractionModel::Status::Failure);
-
-            // TODO: Right now all but the first AddStatus call fail, so this
-            // test is not really testing what it should.
-            for (size_t i = 0; i < 3; ++i)
-            {
-                (void) apCommandObj->FallibleAddStatus(aCommandPath, Protocols::InteractionModel::Status::Failure);
-            }
-        }
-        else if (responseDirective == kSendSuccessStatusCodeWithClusterStatus)
-        {
-            apCommandObj->AddStatus(
-                aCommandPath, Protocols::InteractionModel::ClusterStatusCode::ClusterSpecificSuccess(kTestSuccessClusterStatus));
-        }
-        else if (responseDirective == kSendErrorWithClusterStatus)
-        {
-            apCommandObj->AddStatus(
-                aCommandPath, Protocols::InteractionModel::ClusterStatusCode::ClusterSpecificFailure(kTestFailureClusterStatus));
-        }
-        else if (responseDirective == kAsync)
-        {
-            asyncHandle = apCommandObj;
-        }
-    }
-}
-
-InteractionModel::Status ServerClusterCommandExists(const ConcreteCommandPath & aCommandPath)
-{
-    // Mock cluster catalog, only support commands on one cluster on one endpoint.
-    using InteractionModel::Status;
-
-    if (aCommandPath.mEndpointId != kTestEndpointId)
-    {
-        return Status::UnsupportedEndpoint;
-    }
-
-    if (aCommandPath.mClusterId != Clusters::UnitTesting::Id)
-    {
-        return Status::UnsupportedCluster;
-    }
-
-    return Status::Success;
-}
-} // namespace app
-} // namespace chip
 
 namespace {
 
@@ -256,7 +125,7 @@ TEST_F(TestCommands, TestDataResponse)
     // not safe to do so.
     auto onFailureCb = [&onFailureWasCalled](CHIP_ERROR aError) { onFailureWasCalled = true; };
 
-    responseDirective = kSendDataResponse;
+    gCommandResponseDirective = CommandResponseDirective::kSendDataResponse;
 
     chip::Controller::InvokeCommandRequest(&mpContext->GetExchangeManager(), sessionHandle, kTestEndpointId, request, onSuccessCb,
                                            onFailureCb);
@@ -294,7 +163,7 @@ TEST_F(TestCommands, TestSuccessNoDataResponse)
     // not safe to do so.
     auto onFailureCb = [&onFailureWasCalled](CHIP_ERROR aError) { onFailureWasCalled = true; };
 
-    responseDirective = kSendSuccessStatusCode;
+    gCommandResponseDirective = CommandResponseDirective::kSendSuccessStatusCode;
 
     chip::Controller::InvokeCommandRequest(&mpContext->GetExchangeManager(), sessionHandle, kTestEndpointId, request, onSuccessCb,
                                            onFailureCb);
@@ -332,7 +201,7 @@ TEST_F(TestCommands, TestMultipleSuccessNoDataResponses)
     // not safe to do so.
     auto onFailureCb = [&failureCalls](CHIP_ERROR aError) { ++failureCalls; };
 
-    responseDirective = kSendMultipleSuccessStatusCodes;
+    gCommandResponseDirective = CommandResponseDirective::kSendMultipleSuccessStatusCodes;
 
     Controller::InvokeCommandRequest(&mpContext->GetExchangeManager(), sessionHandle, kTestEndpointId, request, onSuccessCb,
                                      onFailureCb);
@@ -371,7 +240,7 @@ TEST_F(TestCommands, TestAsyncResponse)
     // not safe to do so.
     auto onFailureCb = [&onFailureWasCalled](CHIP_ERROR aError) { onFailureWasCalled = true; };
 
-    responseDirective = kAsync;
+    gCommandResponseDirective = CommandResponseDirective::kAsync;
 
     chip::Controller::InvokeCommandRequest(&mpContext->GetExchangeManager(), sessionHandle, kTestEndpointId, request, onSuccessCb,
                                            onFailureCb);
@@ -381,12 +250,12 @@ TEST_F(TestCommands, TestAsyncResponse)
     EXPECT_TRUE(!onSuccessWasCalled && !onFailureWasCalled && !statusCheck);
     EXPECT_EQ(mpContext->GetExchangeManager().GetNumActiveExchanges(), 2u);
 
-    CommandHandler * commandHandle = asyncHandle.Get();
+    CommandHandler * commandHandle = gAsyncCommandHandle.Get();
     ASSERT_NE(commandHandle, nullptr);
 
     commandHandle->AddStatus(ConcreteCommandPath(kTestEndpointId, request.GetClusterId(), request.GetCommandId()),
                              Protocols::InteractionModel::Status::Success);
-    asyncHandle.Release();
+    gAsyncCommandHandle.Release();
 
     mpContext->DrainAndServiceIO();
 
@@ -416,7 +285,7 @@ TEST_F(TestCommands, TestFailure)
         onFailureWasCalled = true;
     };
 
-    responseDirective = kSendError;
+    gCommandResponseDirective = CommandResponseDirective::kSendError;
 
     chip::Controller::InvokeCommandRequest(&mpContext->GetExchangeManager(), sessionHandle, kTestEndpointId, request, onSuccessCb,
                                            onFailureCb);
@@ -454,7 +323,7 @@ TEST_F(TestCommands, TestMultipleFailures)
         ++failureCalls;
     };
 
-    responseDirective = kSendMultipleErrors;
+    gCommandResponseDirective = CommandResponseDirective::kSendMultipleErrors;
 
     Controller::InvokeCommandRequest(&mpContext->GetExchangeManager(), sessionHandle, kTestEndpointId, request, onSuccessCb,
                                      onFailureCb);
@@ -494,7 +363,7 @@ TEST_F(TestCommands, TestSuccessNoDataResponseWithClusterStatus)
     // not safe to do so.
     auto onFailureCb = [&onFailureWasCalled](CHIP_ERROR aError) { onFailureWasCalled = true; };
 
-    responseDirective = kSendSuccessStatusCodeWithClusterStatus;
+    gCommandResponseDirective = CommandResponseDirective::kSendSuccessStatusCodeWithClusterStatus;
 
     chip::Controller::InvokeCommandRequest(&mpContext->GetExchangeManager(), sessionHandle, kTestEndpointId, request, onSuccessCb,
                                            onFailureCb);
@@ -533,7 +402,7 @@ TEST_F(TestCommands, TestFailureWithClusterStatus)
         onFailureWasCalled = true;
     };
 
-    responseDirective = kSendErrorWithClusterStatus;
+    gCommandResponseDirective = CommandResponseDirective::kSendErrorWithClusterStatus;
 
     chip::Controller::InvokeCommandRequest(&mpContext->GetExchangeManager(), sessionHandle, kTestEndpointId, request, onSuccessCb,
                                            onFailureCb);
