+++ src/controller/tests/data_model/TestRead.cpp
@@ -18,8 +18,8 @@
 
 #include <gtest/gtest.h>
 
-#include "system/SystemClock.h"
-#include "transport/SecureSession.h"
+#include "DataModelFixtures.h"
+
 #include <app-common/zap-generated/cluster-objects.h>
 #include <app/ClusterStateCache.h>
 #include <app/ConcreteAttributePath.h>
@@ -33,228 +33,17 @@
 #include <lib/support/logging/CHIPLogging.h>
 #include <messaging/tests/MessagingContext.h>
 #include <protocols/interaction_model/Constants.h>
+#include <system/SystemClock.h>
+#include <transport/SecureSession.h>
 
 using TestContext = chip::Test::AppContext;
 
 using namespace chip;
 using namespace chip::app;
 using namespace chip::app::Clusters;
+using namespace chip::app::DataModelTests;
 using namespace chip::Protocols;
-
-namespace {
-
-constexpr EndpointId kTestEndpointId        = 1;
-constexpr DataVersion kDataVersion          = 5;
-constexpr AttributeId kPerpetualAttributeid = chip::Test::MockAttributeId(1);
-constexpr ClusterId kPerpetualClusterId     = chip::Test::MockClusterId(2);
-bool expectedAttribute1                     = true;
-int16_t expectedAttribute2                  = 42;
-uint64_t expectedAttribute3                 = 0xdeadbeef0000cafe;
-uint8_t expectedAttribute4[256]             = {
-    0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0xa, 0xb, 0xc, 0xd, 0xe, 0xf, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0xa, 0xb, 0xc, 0xd, 0xe, 0xf,
-    0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0xa, 0xb, 0xc, 0xd, 0xe, 0xf, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0xa, 0xb, 0xc, 0xd, 0xe, 0xf,
-    0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0xa, 0xb, 0xc, 0xd, 0xe, 0xf, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0xa, 0xb, 0xc, 0xd, 0xe, 0xf,
-    0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0xa, 0xb, 0xc, 0xd, 0xe, 0xf, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0xa, 0xb, 0xc, 0xd, 0xe, 0xf,
-    0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0xa, 0xb, 0xc, 0xd, 0xe, 0xf, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0xa, 0xb, 0xc, 0xd, 0xe, 0xf,
-    0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0xa, 0xb, 0xc, 0xd, 0xe, 0xf, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0xa, 0xb, 0xc, 0xd, 0xe, 0xf,
-    0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0xa, 0xb, 0xc, 0xd, 0xe, 0xf, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0xa, 0xb, 0xc, 0xd, 0xe, 0xf,
-    0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0xa, 0xb, 0xc, 0xd, 0xe, 0xf, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0xa, 0xb, 0xc, 0xd, 0xe, 0xf,
-};
-
-enum ResponseDirective
-{
-    kSendDataResponse,
-    kSendManyDataResponses,          // Many data blocks, for a single concrete path
-                                     // read, simulating a malicious server.
-    kSendManyDataResponsesWrongPath, // Many data blocks, all using the wrong
-                                     // path, for a single concrete path
-                                     // read, simulating a malicious server.
-    kSendDataError,
-    kSendTwoDataErrors, // Multiple errors, for a single concrete path,
-                        // simulating a malicious server.
-};
-
-ResponseDirective responseDirective;
-
-// Number of reads of Clusters::UnitTesting::Attributes::Int16u that we have observed.
-// Every read will increment this count by 1 and return the new value.
-uint16_t totalReadCount = 0;
-
-bool isLitIcd = false;
-
-} // namespace
-
-namespace chip {
-namespace app {
-CHIP_ERROR ReadSingleClusterData(const Access::SubjectDescriptor & aSubjectDescriptor, bool aIsFabricFiltered,
-                                 const ConcreteReadAttributePath & aPath, AttributeReportIBs::Builder & aAttributeReports,
-                                 AttributeEncodeState * apEncoderState)
-{
-    if (aPath.mEndpointId >= chip::Test::kMockEndpointMin)
-    {
-        return chip::Test::ReadSingleMockClusterData(aSubjectDescriptor.fabricIndex, aPath, aAttributeReports, apEncoderState);
-    }
-
-    if (responseDirective == kSendManyDataResponses || responseDirective == kSendManyDataResponsesWrongPath)
-    {
-        if (aPath.mClusterId != Clusters::UnitTesting::Id || aPath.mAttributeId != Clusters::UnitTesting::Attributes::Boolean::Id)
-        {
-            return CHIP_ERROR_INCORRECT_STATE;
-        }
-
-        for (size_t i = 0; i < 4; ++i)
-        {
-            ConcreteAttributePath path(aPath);
-            // Use an incorrect attribute id for some of the responses.
-            path.mAttributeId =
-                static_cast<AttributeId>(path.mAttributeId + (i / 2) + (responseDirective == kSendManyDataResponsesWrongPath));
-            AttributeEncodeState state(apEncoderState);
-            AttributeValueEncoder valueEncoder(aAttributeReports, aSubjectDescriptor, path, kDataVersion, aIsFabricFiltered, state);
-            ReturnErrorOnFailure(valueEncoder.Encode(true));
-        }
-
-        return CHIP_NO_ERROR;
-    }
-
-    if (responseDirective == kSendDataResponse)
-    {
-        if (aPath.mClusterId == app::Clusters::UnitTesting::Id &&
-            aPath.mAttributeId == app::Clusters::UnitTesting::Attributes::ListFabricScoped::Id)
-        {
-            AttributeEncodeState state(apEncoderState);
-            AttributeValueEncoder valueEncoder(aAttributeReports, aSubjectDescriptor, aPath, kDataVersion, aIsFabricFiltered,
-                                               state);
-
-            return valueEncoder.EncodeList([aSubjectDescriptor](const auto & encoder) -> CHIP_ERROR {
-                app::Clusters::UnitTesting::Structs::TestFabricScoped::Type val;
-                val.fabricIndex = aSubjectDescriptor.fabricIndex;
-                ReturnErrorOnFailure(encoder.Encode(val));
-                val.fabricIndex = (val.fabricIndex == 1) ? 2 : 1;
-                ReturnErrorOnFailure(encoder.Encode(val));
-                return CHIP_NO_ERROR;
-            });
-        }
-        if (aPath.mClusterId == app::Clusters::UnitTesting::Id &&
-            aPath.mAttributeId == app::Clusters::UnitTesting::Attributes::Int16u::Id)
-        {
-            AttributeEncodeState state(apEncoderState);
-            AttributeValueEncoder valueEncoder(aAttributeReports, aSubjectDescriptor, aPath, kDataVersion, aIsFabricFiltered,
-                                               state);
-
-            return valueEncoder.Encode(++totalReadCount);
-        }
-        if (aPath.mClusterId == kPerpetualClusterId ||
-            (aPath.mClusterId == app::Clusters::UnitTesting::Id && aPath.mAttributeId == kPerpetualAttributeid))
-        {
-            AttributeEncodeState state;
-            AttributeValueEncoder valueEncoder(aAttributeReports, aSubjectDescriptor, aPath, kDataVersion, aIsFabricFiltered,
-                                               state);
-
-            CHIP_ERROR err = valueEncoder.EncodeList([](const auto & encoder) -> CHIP_ERROR {
-                encoder.Encode(static_cast<uint8_t>(1));
-                return CHIP_ERROR_NO_MEMORY;
-            });
-
-            if (err != CHIP_NO_ERROR)
-            {
-                // If the err is not CHIP_NO_ERROR, means the encoding was aborted, then the valueEncoder may save its state.
-                // The state is used by list chunking feature for now.
-                if (apEncoderState != nullptr)
-                {
-                    *apEncoderState = valueEncoder.GetState();
-                }
-                return err;
-            }
-        }
-        if (aPath.mClusterId == app::Clusters::IcdManagement::Id &&
-            aPath.mAttributeId == app::Clusters::IcdManagement::Attributes::OperatingMode::Id)
-        {
-            AttributeEncodeState state(apEncoderState);
-            AttributeValueEncoder valueEncoder(aAttributeReports, aSubjectDescriptor, aPath, kDataVersion, aIsFabricFiltered,
-                                               state);
-
-            return valueEncoder.Encode(isLitIcd ? Clusters::IcdManagement::OperatingModeEnum::kLit
-                                                : Clusters::IcdManagement::OperatingModeEnum::kSit);
-        }
-
-        AttributeReportIB::Builder & attributeReport = aAttributeReports.CreateAttributeReport();
-        ReturnErrorOnFailure(aAttributeReports.GetError());
-        AttributeDataIB::Builder & attributeData = attributeReport.CreateAttributeData();
-        ReturnErrorOnFailure(attributeReport.GetError());
-        Clusters::UnitTesting::Attributes::ListStructOctetString::TypeInfo::Type value;
-        Clusters::UnitTesting::Structs::TestListStructOctet::Type valueBuf[4];
-
-        value = valueBuf;
-
-        uint8_t i = 0;
-        for (auto & item : valueBuf)
-        {
-            item.member1 = i;
-            i++;
-        }
-
-        attributeData.DataVersion(kDataVersion);
-        ReturnErrorOnFailure(attributeData.GetError());
-        AttributePathIB::Builder & attributePath = attributeData.CreatePath();
-        attributePath.Endpoint(aPath.mEndpointId).Cluster(aPath.mClusterId).Attribute(aPath.mAttributeId).EndOfAttributePathIB();
-        ReturnErrorOnFailure(attributePath.GetError());
-
-        ReturnErrorOnFailure(DataModel::Encode(*(attributeData.GetWriter()), TLV::ContextTag(AttributeDataIB::Tag::kData), value));
-        ReturnErrorOnFailure(attributeData.EndOfAttributeDataIB());
-        return attributeReport.EndOfAttributeReportIB();
-    }
-
-    for (size_t i = 0; i < (responseDirective == kSendTwoDataErrors ? 2 : 1); ++i)
-    {
-        AttributeReportIB::Builder & attributeReport = aAttributeReports.CreateAttributeReport();
-        ReturnErrorOnFailure(aAttributeReports.GetError());
-        AttributeStatusIB::Builder & attributeStatus = attributeReport.CreateAttributeStatus();
-        AttributePathIB::Builder & attributePath     = attributeStatus.CreatePath();
-        attributePath.Endpoint(aPath.mEndpointId).Cluster(aPath.mClusterId).Attribute(aPath.mAttributeId).EndOfAttributePathIB();
-        ReturnErrorOnFailure(attributePath.GetError());
-
-        StatusIB::Builder & errorStatus = attributeStatus.CreateErrorStatus();
-        ReturnErrorOnFailure(attributeStatus.GetError());
-        errorStatus.EncodeStatusIB(StatusIB(Protocols::InteractionModel::Status::Busy));
-        attributeStatus.EndOfAttributeStatusIB();
-        ReturnErrorOnFailure(attributeStatus.GetError());
-        ReturnErrorOnFailure(attributeReport.EndOfAttributeReportIB());
-    }
-
-    return CHIP_NO_ERROR;
-}
-
-bool IsClusterDataVersionEqual(const app::ConcreteClusterPath & aConcreteClusterPath, DataVersion aRequiredVersion)
-{
-    if (aRequiredVersion == kDataVersion)
-    {
-        return true;
-    }
-    if (Test::GetVersion() == aRequiredVersion)
-    {
-        return true;
-    }
-
-    return false;
-}
-
-bool IsDeviceTypeOnEndpoint(DeviceTypeId deviceType, EndpointId endpoint)
-{
-    return false;
-}
-
-bool ConcreteAttributePathExists(const ConcreteAttributePath & aPath)
-{
-    return true;
-}
-
-Protocols::InteractionModel::Status CheckEventSupportStatus(const ConcreteEventPath & aPath)
-{
-    return Protocols::InteractionModel::Status::Success;
-}
-
-} // namespace app
-} // namespace chip
+using namespace chip::Test;
 
 namespace {
 
@@ -388,7 +177,7 @@ TEST_F(TestRead, TestReadAttributeResponse)
     auto sessionHandle      = mpContext->GetSessionBobToAlice();
     bool onSuccessCbInvoked = false, onFailureCbInvoked = false;
 
-    responseDirective = kSendDataResponse;
+    gReadResponseDirective = ReadResponseDirective::kSendDataResponse;
 
     // Passing of stack variables by reference is only safe because of synchronous completion of the interaction. Otherwise, it's
     // not safe to do so.
@@ -428,8 +217,8 @@ TEST_F(TestRead, TestReadAttributeResponse)
 // `EXPECT_TRUE(version1.HasValue() && (version1.Value() == 0))`.
 TEST_F(TestRead, TestReadSubscribeAttributeResponseWithVersionOnlyCache)
 {
-    CHIP_ERROR err    = CHIP_NO_ERROR;
-    responseDirective = kSendDataResponse;
+    CHIP_ERROR err         = CHIP_NO_ERROR;
+    gReadResponseDirective = ReadResponseDirective::kSendDataResponse;
 
     MockInteractionModelApp delegate;
     chip::app::ClusterStateCache cache(delegate, Optional<EventNumber>::Missing(), false /*cachedData*/);
@@ -500,8 +289,8 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithVersionOnlyCache)
 
 TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
 {
-    CHIP_ERROR err    = CHIP_NO_ERROR;
-    responseDirective = kSendDataResponse;
+    CHIP_ERROR err         = CHIP_NO_ERROR;
+    gReadResponseDirective = ReadResponseDirective::kSendDataResponse;
 
     MockInteractionModelApp delegate;
     chip::app::ClusterStateCache cache(delegate);
@@ -604,7 +393,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             bool receivedAttribute1;
             reader.Get(receivedAttribute1);
-            EXPECT_EQ(receivedAttribute1, expectedAttribute1);
+            EXPECT_EQ(receivedAttribute1, mockAttribute1);
         }
 
         {
@@ -614,7 +403,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             int16_t receivedAttribute2;
             reader.Get(receivedAttribute2);
-            EXPECT_EQ(receivedAttribute2, expectedAttribute2);
+            EXPECT_EQ(receivedAttribute2, mockAttribute2);
         }
 
         {
@@ -624,7 +413,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             int16_t receivedAttribute2;
             reader.Get(receivedAttribute2);
-            EXPECT_EQ(receivedAttribute2, expectedAttribute2);
+            EXPECT_EQ(receivedAttribute2, mockAttribute2);
         }
 
         delegate.mNumAttributeResponse = 0;
@@ -682,7 +471,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             int16_t receivedAttribute2;
             reader.Get(receivedAttribute2);
-            EXPECT_EQ(receivedAttribute2, expectedAttribute2);
+            EXPECT_EQ(receivedAttribute2, mockAttribute2);
         }
 
         {
@@ -692,7 +481,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             int16_t receivedAttribute2;
             reader.Get(receivedAttribute2);
-            EXPECT_EQ(receivedAttribute2, expectedAttribute2);
+            EXPECT_EQ(receivedAttribute2, mockAttribute2);
         }
         delegate.mNumAttributeResponse = 0;
     }
@@ -737,7 +526,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             bool receivedAttribute1;
             reader.Get(receivedAttribute1);
-            EXPECT_EQ(receivedAttribute1, expectedAttribute1);
+            EXPECT_EQ(receivedAttribute1, mockAttribute1);
         }
 
         {
@@ -747,7 +536,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             int16_t receivedAttribute2;
             reader.Get(receivedAttribute2);
-            EXPECT_EQ(receivedAttribute2, expectedAttribute2);
+            EXPECT_EQ(receivedAttribute2, mockAttribute2);
         }
 
         {
@@ -757,7 +546,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             uint64_t receivedAttribute3;
             reader.Get(receivedAttribute3);
-            EXPECT_EQ(receivedAttribute3, expectedAttribute3);
+            EXPECT_EQ(receivedAttribute3, mockAttribute3);
         }
 
         {
@@ -767,7 +556,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             int16_t receivedAttribute2;
             reader.Get(receivedAttribute2);
-            EXPECT_EQ(receivedAttribute2, expectedAttribute2);
+            EXPECT_EQ(receivedAttribute2, mockAttribute2);
         }
         delegate.mNumAttributeResponse = 0;
     }
@@ -816,7 +605,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             bool receivedAttribute1;
             reader.Get(receivedAttribute1);
-            EXPECT_EQ(receivedAttribute1, expectedAttribute1);
+            EXPECT_EQ(receivedAttribute1, mockAttribute1);
         }
 
         {
@@ -826,7 +615,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             int16_t receivedAttribute2;
             reader.Get(receivedAttribute2);
-            EXPECT_EQ(receivedAttribute2, expectedAttribute2);
+            EXPECT_EQ(receivedAttribute2, mockAttribute2);
         }
 
         {
@@ -836,7 +625,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             int16_t receivedAttribute2;
             reader.Get(receivedAttribute2);
-            EXPECT_EQ(receivedAttribute2, expectedAttribute2);
+            EXPECT_EQ(receivedAttribute2, mockAttribute2);
         }
         delegate.mNumAttributeResponse = 0;
     }
@@ -880,7 +669,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             bool receivedAttribute1;
             reader.Get(receivedAttribute1);
-            EXPECT_EQ(receivedAttribute1, expectedAttribute1);
+            EXPECT_EQ(receivedAttribute1, mockAttribute1);
         }
 
         {
@@ -890,7 +679,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             int16_t receivedAttribute2;
             reader.Get(receivedAttribute2);
-            EXPECT_EQ(receivedAttribute2, expectedAttribute2);
+            EXPECT_EQ(receivedAttribute2, mockAttribute2);
         }
 
         {
@@ -900,7 +689,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             uint64_t receivedAttribute3;
             reader.Get(receivedAttribute3);
-            EXPECT_EQ(receivedAttribute3, expectedAttribute3);
+            EXPECT_EQ(receivedAttribute3, mockAttribute3);
         }
 
         {
@@ -910,7 +699,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             int16_t receivedAttribute2;
             reader.Get(receivedAttribute2);
-            EXPECT_EQ(receivedAttribute2, expectedAttribute2);
+            EXPECT_EQ(receivedAttribute2, mockAttribute2);
         }
         delegate.mNumAttributeResponse = 0;
     }
@@ -962,7 +751,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             bool receivedAttribute1;
             reader.Get(receivedAttribute1);
-            EXPECT_EQ(receivedAttribute1, expectedAttribute1);
+            EXPECT_EQ(receivedAttribute1, mockAttribute1);
         }
 
         {
@@ -972,7 +761,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             int16_t receivedAttribute2;
             reader.Get(receivedAttribute2);
-            EXPECT_EQ(receivedAttribute2, expectedAttribute2);
+            EXPECT_EQ(receivedAttribute2, mockAttribute2);
         }
 
         {
@@ -982,7 +771,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             int16_t receivedAttribute2;
             reader.Get(receivedAttribute2);
-            EXPECT_EQ(receivedAttribute2, expectedAttribute2);
+            EXPECT_EQ(receivedAttribute2, mockAttribute2);
         }
         delegate.mNumAttributeResponse = 0;
     }
@@ -1031,7 +820,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             bool receivedAttribute1;
             reader.Get(receivedAttribute1);
-            EXPECT_EQ(receivedAttribute1, expectedAttribute1);
+            EXPECT_EQ(receivedAttribute1, mockAttribute1);
         }
 
         {
@@ -1041,7 +830,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             int16_t receivedAttribute2;
             reader.Get(receivedAttribute2);
-            EXPECT_EQ(receivedAttribute2, expectedAttribute2);
+            EXPECT_EQ(receivedAttribute2, mockAttribute2);
         }
 
         {
@@ -1051,7 +840,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             int16_t receivedAttribute2;
             reader.Get(receivedAttribute2);
-            EXPECT_EQ(receivedAttribute2, expectedAttribute2);
+            EXPECT_EQ(receivedAttribute2, mockAttribute2);
         }
         delegate.mNumAttributeResponse = 0;
     }
@@ -1095,7 +884,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             bool receivedAttribute1;
             reader.Get(receivedAttribute1);
-            EXPECT_EQ(receivedAttribute1, expectedAttribute1);
+            EXPECT_EQ(receivedAttribute1, mockAttribute1);
         }
 
         {
@@ -1105,7 +894,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             int16_t receivedAttribute2;
             reader.Get(receivedAttribute2);
-            EXPECT_EQ(receivedAttribute2, expectedAttribute2);
+            EXPECT_EQ(receivedAttribute2, mockAttribute2);
         }
 
         {
@@ -1115,7 +904,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             uint64_t receivedAttribute3;
             reader.Get(receivedAttribute3);
-            EXPECT_EQ(receivedAttribute3, expectedAttribute3);
+            EXPECT_EQ(receivedAttribute3, mockAttribute3);
         }
 
         {
@@ -1125,7 +914,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             int16_t receivedAttribute2;
             reader.Get(receivedAttribute2);
-            EXPECT_EQ(receivedAttribute2, expectedAttribute2);
+            EXPECT_EQ(receivedAttribute2, mockAttribute2);
         }
         delegate.mNumAttributeResponse = 0;
     }
@@ -1178,7 +967,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             bool receivedAttribute1;
             reader.Get(receivedAttribute1);
-            EXPECT_EQ(receivedAttribute1, expectedAttribute1);
+            EXPECT_EQ(receivedAttribute1, mockAttribute1);
         }
 
         {
@@ -1188,7 +977,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             int16_t receivedAttribute2;
             reader.Get(receivedAttribute2);
-            EXPECT_EQ(receivedAttribute2, expectedAttribute2);
+            EXPECT_EQ(receivedAttribute2, mockAttribute2);
         }
 
         {
@@ -1198,7 +987,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             uint64_t receivedAttribute3;
             reader.Get(receivedAttribute3);
-            EXPECT_EQ(receivedAttribute3, expectedAttribute3);
+            EXPECT_EQ(receivedAttribute3, mockAttribute3);
         }
 
         {
@@ -1208,7 +997,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             int16_t receivedAttribute2;
             reader.Get(receivedAttribute2);
-            EXPECT_EQ(receivedAttribute2, expectedAttribute2);
+            EXPECT_EQ(receivedAttribute2, mockAttribute2);
         }
         delegate.mNumAttributeResponse             = 0;
         readPrepareParams.mpEventPathParamsList    = nullptr;
@@ -1267,7 +1056,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             bool receivedAttribute1;
             reader.Get(receivedAttribute1);
-            EXPECT_EQ(receivedAttribute1, expectedAttribute1);
+            EXPECT_EQ(receivedAttribute1, mockAttribute1);
         }
 
         {
@@ -1277,7 +1066,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             bool receivedAttribute1;
             reader.Get(receivedAttribute1);
-            EXPECT_EQ(receivedAttribute1, expectedAttribute1);
+            EXPECT_EQ(receivedAttribute1, mockAttribute1);
         }
 
         {
@@ -1287,7 +1076,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             int16_t receivedAttribute2;
             reader.Get(receivedAttribute2);
-            EXPECT_EQ(receivedAttribute2, expectedAttribute2);
+            EXPECT_EQ(receivedAttribute2, mockAttribute2);
         }
 
         {
@@ -1297,7 +1086,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             uint64_t receivedAttribute3;
             reader.Get(receivedAttribute3);
-            EXPECT_EQ(receivedAttribute3, expectedAttribute3);
+            EXPECT_EQ(receivedAttribute3, mockAttribute3);
         }
 
         {
@@ -1307,7 +1096,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             bool receivedAttribute1;
             reader.Get(receivedAttribute1);
-            EXPECT_EQ(receivedAttribute1, expectedAttribute1);
+            EXPECT_EQ(receivedAttribute1, mockAttribute1);
         }
         {
             app::ConcreteAttributePath attributePath(chip::Test::kMockEndpoint2, chip::Test::MockClusterId(2),
@@ -1316,7 +1105,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             int16_t receivedAttribute2;
             reader.Get(receivedAttribute2);
-            EXPECT_EQ(receivedAttribute2, expectedAttribute2);
+            EXPECT_EQ(receivedAttribute2, mockAttribute2);
         }
 
         delegate.mNumAttributeResponse = 0;
@@ -1378,7 +1167,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             bool receivedAttribute1;
             reader.Get(receivedAttribute1);
-            EXPECT_EQ(receivedAttribute1, expectedAttribute1);
+            EXPECT_EQ(receivedAttribute1, mockAttribute1);
         }
 
         {
@@ -1388,7 +1177,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             bool receivedAttribute1;
             reader.Get(receivedAttribute1);
-            EXPECT_EQ(receivedAttribute1, expectedAttribute1);
+            EXPECT_EQ(receivedAttribute1, mockAttribute1);
         }
 
         {
@@ -1398,7 +1187,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             int16_t receivedAttribute2;
             reader.Get(receivedAttribute2);
-            EXPECT_EQ(receivedAttribute2, expectedAttribute2);
+            EXPECT_EQ(receivedAttribute2, mockAttribute2);
         }
 
         {
@@ -1408,7 +1197,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             uint64_t receivedAttribute3;
             reader.Get(receivedAttribute3);
-            EXPECT_EQ(receivedAttribute3, expectedAttribute3);
+            EXPECT_EQ(receivedAttribute3, mockAttribute3);
         }
 
         {
@@ -1418,7 +1207,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             bool receivedAttribute1;
             reader.Get(receivedAttribute1);
-            EXPECT_EQ(receivedAttribute1, expectedAttribute1);
+            EXPECT_EQ(receivedAttribute1, mockAttribute1);
         }
 
         {
@@ -1428,7 +1217,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             int16_t receivedAttribute2;
             reader.Get(receivedAttribute2);
-            EXPECT_EQ(receivedAttribute2, expectedAttribute2);
+            EXPECT_EQ(receivedAttribute2, mockAttribute2);
         }
         delegate.mNumAttributeResponse             = 0;
         readPrepareParams.mpEventPathParamsList    = nullptr;
@@ -1467,7 +1256,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             bool receivedAttribute1;
             reader.Get(receivedAttribute1);
-            EXPECT_EQ(receivedAttribute1, expectedAttribute1);
+            EXPECT_EQ(receivedAttribute1, mockAttribute1);
         }
 
         {
@@ -1477,7 +1266,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             int16_t receivedAttribute2;
             reader.Get(receivedAttribute2);
-            EXPECT_EQ(receivedAttribute2, expectedAttribute2);
+            EXPECT_EQ(receivedAttribute2, mockAttribute2);
         }
 
         {
@@ -1487,7 +1276,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             uint64_t receivedAttribute3;
             reader.Get(receivedAttribute3);
-            EXPECT_EQ(receivedAttribute3, expectedAttribute3);
+            EXPECT_EQ(receivedAttribute3, mockAttribute3);
         }
 
         {
@@ -1497,7 +1286,7 @@ TEST_F(TestRead, TestReadSubscribeAttributeResponseWithCache)
             EXPECT_EQ(cache.Get(attributePath, reader), CHIP_NO_ERROR);
             uint8_t receivedAttribute4[256];
             reader.GetBytes(receivedAttribute4, 256);
-            EXPECT_TRUE(memcmp(receivedAttribute4, expectedAttribute4, 256));
+            EXPECT_TRUE(memcmp(receivedAttribute4, mockAttribute4, 256));
         }
         delegate.mNumAttributeResponse = 0;
     }
@@ -1544,7 +1333,7 @@ TEST_F(TestRead, TestReadAttributeError)
     auto sessionHandle      = mpContext->GetSessionBobToAlice();
     bool onSuccessCbInvoked = false, onFailureCbInvoked = false;
 
-    responseDirective = kSendDataError;
+    gReadResponseDirective = ReadResponseDirective::kSendDataError;
 
     // Passing of stack variables by reference is only safe because of synchronous completion of the interaction. Otherwise, it's
     // not safe to do so.
@@ -1575,7 +1364,7 @@ TEST_F(TestRead, TestReadAttributeTimeout)
     auto sessionHandle      = mpContext->GetSessionBobToAlice();
     bool onSuccessCbInvoked = false, onFailureCbInvoked = false;
 
-    responseDirective = kSendDataError;
+    gReadResponseDirective = ReadResponseDirective::kSendDataError;
 
     // Passing of stack variables by reference is only safe because of synchronous completion of the interaction. Otherwise, it's
     // not safe to do so.
@@ -1834,7 +1623,7 @@ TEST_F(TestRead, TestReadHandler_MultipleSubscriptions)
     uint32_t numSuccessCalls                 = 0;
     uint32_t numSubscriptionEstablishedCalls = 0;
 
-    responseDirective = kSendDataResponse;
+    gReadResponseDirective = ReadResponseDirective::kSendDataResponse;
 
     // Passing of stack variables by reference is only safe because of synchronous completion of the interaction. Otherwise, it's
     // not safe to do so.
@@ -1902,7 +1691,7 @@ TEST_F(TestRead, TestReadHandler_SubscriptionAppRejection)
     uint32_t numFailureCalls                 = 0;
     uint32_t numSubscriptionEstablishedCalls = 0;
 
-    responseDirective = kSendDataResponse;
+    gReadResponseDirective = ReadResponseDirective::kSendDataResponse;
 
     // Passing of stack variables by reference is only safe because of synchronous completion of the interaction. Otherwise, it's
     // not safe to do so.
@@ -1968,7 +1757,7 @@ TEST_F(TestRead, TestReadHandler_SubscriptionReportingIntervalsTest1)
     uint32_t numFailureCalls                 = 0;
     uint32_t numSubscriptionEstablishedCalls = 0;
 
-    responseDirective = kSendDataResponse;
+    gReadResponseDirective = ReadResponseDirective::kSendDataResponse;
 
     // Passing of stack variables by reference is only safe because of synchronous completion of the interaction. Otherwise, it's
     // not safe to do so.
@@ -2043,7 +1832,7 @@ TEST_F(TestRead, TestReadHandler_SubscriptionReportingIntervalsTest2)
     uint32_t numFailureCalls                 = 0;
     uint32_t numSubscriptionEstablishedCalls = 0;
 
-    responseDirective = kSendDataResponse;
+    gReadResponseDirective = ReadResponseDirective::kSendDataResponse;
 
     // Passing of stack variables by reference is only safe because of synchronous completion of the interaction. Otherwise, it's
     // not safe to do so.
@@ -2118,7 +1907,7 @@ TEST_F(TestRead, TestReadHandler_SubscriptionReportingIntervalsTest3)
     uint32_t numFailureCalls                 = 0;
     uint32_t numSubscriptionEstablishedCalls = 0;
 
-    responseDirective = kSendDataResponse;
+    gReadResponseDirective = ReadResponseDirective::kSendDataResponse;
 
     // Passing of stack variables by reference is only safe because of synchronous completion of the interaction. Otherwise, it's
     // not safe to do so.
@@ -2195,7 +1984,7 @@ TEST_F(TestRead, TestReadHandler_SubscriptionReportingIntervalsTest4)
     uint32_t numFailureCalls                 = 0;
     uint32_t numSubscriptionEstablishedCalls = 0;
 
-    responseDirective = kSendDataResponse;
+    gReadResponseDirective = ReadResponseDirective::kSendDataResponse;
 
     // Passing of stack variables by reference is only safe because of synchronous completion of the interaction. Otherwise, it's
     // not safe to do so.
@@ -2262,7 +2051,7 @@ TEST_F(TestRead, TestReadHandler_SubscriptionReportingIntervalsTest5)
     uint32_t numFailureCalls                 = 0;
     uint32_t numSubscriptionEstablishedCalls = 0;
 
-    responseDirective = kSendDataResponse;
+    gReadResponseDirective = ReadResponseDirective::kSendDataResponse;
 
     // Passing of stack variables by reference is only safe because of synchronous completion of the interaction. Otherwise, it's
     // not safe to do so.
@@ -2337,7 +2126,7 @@ TEST_F(TestRead, TestReadHandler_SubscriptionReportingIntervalsTest6)
     uint32_t numFailureCalls                 = 0;
     uint32_t numSubscriptionEstablishedCalls = 0;
 
-    responseDirective = kSendDataResponse;
+    gReadResponseDirective = ReadResponseDirective::kSendDataResponse;
 
     // Passing of stack variables by reference is only safe because of synchronous completion of the interaction. Otherwise, it's
     // not safe to do so.
@@ -2412,7 +2201,7 @@ TEST_F(TestRead, TestReadHandler_SubscriptionReportingIntervalsTest7)
     uint32_t numFailureCalls                 = 0;
     uint32_t numSubscriptionEstablishedCalls = 0;
 
-    responseDirective = kSendDataResponse;
+    gReadResponseDirective = ReadResponseDirective::kSendDataResponse;
 
     // Passing of stack variables by reference is only safe because of synchronous completion of the interaction. Otherwise, it's
     // not safe to do so.
@@ -2488,7 +2277,7 @@ TEST_F(TestRead, TestReadHandler_SubscriptionReportingIntervalsTest8)
     uint32_t numFailureCalls                 = 0;
     uint32_t numSubscriptionEstablishedCalls = 0;
 
-    responseDirective = kSendDataResponse;
+    gReadResponseDirective = ReadResponseDirective::kSendDataResponse;
 
     // Passing of stack variables by reference is only safe because of synchronous completion of the interaction. Otherwise, it's
     // not safe to do so.
@@ -2551,7 +2340,7 @@ TEST_F(TestRead, TestReadHandler_SubscriptionReportingIntervalsTest9)
     uint32_t numFailureCalls                 = 0;
     uint32_t numSubscriptionEstablishedCalls = 0;
 
-    responseDirective = kSendDataResponse;
+    gReadResponseDirective = ReadResponseDirective::kSendDataResponse;
 
     // Passing of stack variables by reference is only safe because of synchronous completion of the interaction. Otherwise, it's
     // not safe to do so.
@@ -2704,14 +2493,15 @@ TEST_F(TestRead, TestSubscribe_DynamicLITSubscription)
     mpContext->SetMRPMode(chip::Test::MessagingContext::MRPMode::kResponsive);
 
     {
+        gReadResponseDirective = ReadResponseDirective::kSendDataResponse;
+        gIsLitIcd              = false;
+
         TestResubscriptionCallback callback;
         app::ReadClient readClient(app::InteractionModelEngine::GetInstance(), &mpContext->GetExchangeManager(), callback,
                                    app::ReadClient::InteractionType::Subscribe);
 
-        responseDirective                           = kSendDataResponse;
         callback.mScheduleLITResubscribeImmediately = false;
         callback.SetReadClient(&readClient);
-        isLitIcd = false;
 
         app::ReadPrepareParams readPrepareParams(mpContext->GetSessionBobToAlice());
 
@@ -2778,7 +2568,7 @@ TEST_F(TestRead, TestSubscribe_DynamicLITSubscription)
 
         // Part 2. SIT -> LIT
 
-        isLitIcd = true;
+        gIsLitIcd = true;
         {
             app::AttributePathParams path;
             path.mEndpointId  = kRootEndpointId;
@@ -2807,7 +2597,7 @@ TEST_F(TestRead, TestSubscribe_DynamicLITSubscription)
     app::InteractionModelEngine::GetInstance()->ShutdownActiveReads();
     EXPECT_EQ(mpContext->GetExchangeManager().GetNumActiveExchanges(), 0u);
 
-    isLitIcd = false;
+    gIsLitIcd = false;
 }
 
 /**
@@ -2942,7 +2732,7 @@ void TestRead::SubscribeThenReadHelper(TestContext * apCtx, size_t aSubscribeCou
     uint32_t numReadSuccessCalls = 0;
     uint32_t numReadFailureCalls = 0;
 
-    responseDirective = kSendDataResponse;
+    gReadResponseDirective = ReadResponseDirective::kSendDataResponse;
 
     // Passing of stack variables by reference is only safe because of synchronous completion of the interaction. Otherwise, it's
     // not safe to do so.
@@ -2995,9 +2785,9 @@ void TestRead::MultipleReadHelperInternal(TestContext * apCtx, size_t aReadCount
 
     auto sessionHandle = apCtx->GetSessionBobToAlice();
 
-    responseDirective = kSendDataResponse;
+    gReadResponseDirective = ReadResponseDirective::kSendDataResponse;
 
-    uint16_t firstExpectedResponse = totalReadCount + 1;
+    uint16_t firstExpectedResponse = gInt16uTotalReadCount + 1;
 
     auto onFailureCb = [&aNumFailureCalls](const app::ConcreteDataAttributePath * attributePath, CHIP_ERROR aError) {
         aNumFailureCalls++;
@@ -3038,7 +2828,7 @@ TEST_F(TestRead, TestReadHandler_MultipleSubscriptionsWithDataVersionFilter)
     uint32_t numSuccessCalls                 = 0;
     uint32_t numSubscriptionEstablishedCalls = 0;
 
-    responseDirective = kSendDataResponse;
+    gReadResponseDirective = ReadResponseDirective::kSendDataResponse;
 
     // Passing of stack variables by reference is only safe because of synchronous completion of the interaction. Otherwise, it's
     // not safe to do so.
@@ -3099,7 +2889,7 @@ TEST_F(TestRead, TestReadHandlerResourceExhaustion_MultipleReads)
     uint32_t numSuccessCalls = 0;
     uint32_t numFailureCalls = 0;
 
-    responseDirective = kSendDataResponse;
+    gReadResponseDirective = ReadResponseDirective::kSendDataResponse;
 
     // Passing of stack variables by reference is only safe because of synchronous completion of the interaction. Otherwise, it's
     // not safe to do so.
@@ -3150,7 +2940,7 @@ TEST_F(TestRead, TestReadFabricScopedWithoutFabricFilter)
     auto sessionHandle      = mpContext->GetSessionBobToAlice();
     bool onSuccessCbInvoked = false, onFailureCbInvoked = false;
 
-    responseDirective = kSendDataResponse;
+    gReadResponseDirective = ReadResponseDirective::kSendDataResponse;
 
     // Passing of stack variables by reference is only safe because of synchronous completion of the interaction. Otherwise, it's
     // not safe to do so.
@@ -3195,7 +2985,7 @@ TEST_F(TestRead, TestReadFabricScopedWithFabricFilter)
     auto sessionHandle      = mpContext->GetSessionBobToAlice();
     bool onSuccessCbInvoked = false, onFailureCbInvoked = false;
 
-    responseDirective = kSendDataResponse;
+    gReadResponseDirective = ReadResponseDirective::kSendDataResponse;
 
     // Passing of stack variables by reference is only safe because of synchronous completion of the interaction. Otherwise, it's
     // not safe to do so.
@@ -4753,7 +4543,7 @@ TEST_F(TestRead, TestReadAttribute_ManyDataValues)
     size_t successCalls = 0;
     size_t failureCalls = 0;
 
-    responseDirective = kSendManyDataResponses;
+    gReadResponseDirective = ReadResponseDirective::kSendManyDataResponses;
 
     // Passing of stack variables by reference is only safe because of synchronous completion of the interaction. Otherwise, it's
     // not safe to do so.
@@ -4786,7 +4576,7 @@ TEST_F(TestRead, TestReadAttribute_ManyDataValuesWrongPath)
     size_t successCalls = 0;
     size_t failureCalls = 0;
 
-    responseDirective = kSendManyDataResponsesWrongPath;
+    gReadResponseDirective = ReadResponseDirective::kSendManyDataResponsesWrongPath;
 
     // Passing of stack variables by reference is only safe because of synchronous completion of the interaction. Otherwise, it's
     // not safe to do so.
@@ -4819,7 +4609,7 @@ TEST_F(TestRead, TestReadAttribute_ManyErrors)
     size_t successCalls = 0;
     size_t failureCalls = 0;
 
-    responseDirective = kSendTwoDataErrors;
+    gReadResponseDirective = ReadResponseDirective::kSendTwoDataErrors;
 
     // Passing of stack variables by reference is only safe because of synchronous completion of the interaction. Otherwise, it's
     // not safe to do so.
